+++ DaftarADT/ADTPrimitif/point.h
@@ -62,7 +62,7 @@ POINT NextX (POINT P);
 /* Mengirim salinan P dengan absis ditambah satu */              
 POINT NextY (POINT P);
 /* Mengirim salinan P dengan ordinat ditambah satu */
-POINT PlusDelta (POINT P, float deltaX, float deltaY);
+POINT PlusDelta (POINT P, int deltaX, int deltaY);
 /* Mengirim salinan P yang absisnya adalah Absis(P) + deltaX dan ordinatnya adalah Ordinat(P) + deltaY */
 POINT MirrorOf (POINT P, boolean SbX);
 /* Menghasilkan salinan P yang dicerminkan terhadap salah satu sumbu */
